Added a stripped down version of java.util. A few classes (and friends) with fiddly dependencies were removed: Date, Calendar, Properties, ResourceBundle.
I also had to preserve the type of java.util.Iterator because
java.lang.Iterable references it directly and we cant change that. I also ended
up preserving java.util.Enumeration just to simplify a couple of things as
well.

Javarifier chokes with a mysterious error message, so I'll be sorting that out
next.