Some smaller cleanup changes in Uno.java
Wrote a simple shell script in the test folder which generates the uno file needed in the analysis for this test (has to be called from inside the test folder prior to "ant grok"). I would like to integrate it into the build.xml ant file, but will first have to figure out how this works.