JastAdd's nullability inference tool has what I would describe in polite company as a "deficient" algorithm for determining where to place its import statements. It also doesn't support adding multiple imports even though it sometimes emits multiple annotations. In short, fail.
I wrote a little perl script to detect where to place the imports using a more
"sophisticated" approach, and to also detect which annotations are used in the
file so that we can add imports for just what we need and only to files that
need them. Yay perl.

I also added a Raw annotation since it emits that. And discovered that none of
the ElementType enums used in the @Target annotation seem to apply to method
receiver position, but fortunately if you don't put any @Target annotation,
then that means, let an annotation go anywhere. Woo! Fast and loose baby!

I also had to switch to compiling with 1.7 because only 1.7 allows annotations
in method receiver position. Joy. Given that we will give them a precompiled
Nenya library, they should be able to look at 1.7 source and javadocs without
having to actually compile it.